// packages/getVoteData/src/typings.d.ts
export namespace WeiboTypes {
  interface RootObject {
    visible: Visible;
    created_at: string;
    id: number;
    idstr: string;
    mid: string;
    mblogid: string;
    user: User;
    can_edit: boolean;
    textLength: number;
    annotations: Annotation[];
    source: string;
    favorited: boolean;
    rid: string;
    pic_num: number;
    is_paid: boolean;
    pic_bg_new: string;
    mblog_vip_type: number;
    number_display_strategy: Numberdisplaystrategy;
    title_source: Titlesource;
    reposts_count: number;
    comments_count: number;
    attitudes_count: number;
    attitudes_status: number;
    continue_tag: Continuetag;
    isLongText: boolean;
    longText: LongText;
    mlevel: number;
    content_auth: number;
    is_show_bulletin: number;
    comment_manage_info: Commentmanageinfo;
    screen_name_suffix_new: Screennamesuffixnew[];
    share_repost_type: number;
    url_struct: Urlstruct[];
    title: Title;
    mblogtype: number;
    showFeedRepost: boolean;
    showFeedComment: boolean;
    pictureViewerSign: boolean;
    showPictureViewer: boolean;
    rcList: any[];
    can_remark: boolean;
    common_struct: Commonstruct[];
    analysis_extra: string;
    readtimetype: string;
    mixed_count: number;
    is_show_mixed: boolean;
    isSinglePayAudio: boolean;
    text: string;
    text_raw: string;
    region_name: string;
    page_info: Pageinfo;
    ok: number;
  }

  interface Title {
    text: string;
    base_color: number;
    icon_url: string;
  }

  interface Screennamesuffixnew {
    content: string;
    remark?: string;
    color: string;
    dark_color: string;
    type: number;
    icons?: Icon[];
    icons_location?: number;
    truncation: Truncation;
    scheme?: string;
    actionlog?: Actionlog4;
  }

  interface Actionlog4 {
    act_code: number;
    oid: string;
    uicode: null;
    luicode: null;
    fid: null;
    ext: string;
  }

  interface Truncation {
    mode: number;
    keep_end_size?: number;
  }

  interface Icon {
    name?: string;
    url: string;
    scheme: string;
    length?: number;
  }

  interface Commentmanageinfo {
    comment_permission_type: number;
    approval_comment_type: number;
    comment_sort_type: number;
  }

  interface LongText {
    created_at: string;
    appid: number;
    annotations: Annotation[];
    mblog_vip_type: number;
    user: User2;
    weibo_position: number;
    level_new: Levelnew;
    show_attitude_bar: number;
    page_info: Pageinfo;
    url_struct: Urlstruct[];
    common_struct: Commonstruct[];
    content: string;
  }

  interface Commonstruct {
    url: string;
    name: string;
    desc: string;
    img: string;
    type: number;
    btn_show_type: string;
    page_id: string;
    actionlog: Actionlog2;
    buttons: Button[];
    hidden: number;
  }
// ...
  interface Urlstruct {
    url_title: string;
    url_type_pic: string;
    ori_url: string;
    page_id: string;
    short_url: string;
    long_url: string;
    url_type: number | string;
    result: boolean;
    actionlog: Actionlog;
    storage_type: string;
    hide: number;
    object_type: string;
    h5_target_url: string;
    need_save_obj: number;
    log?: string;
  }

  interface Pageinfo {
    type: string;
    page_id: string;
    object_type: string;
    page_desc: string;
    page_title: string;
    page_pic: string;
    type_icon: string;
    page_url: string;
    object_id: string;
    actionlog: Actionlog;
    card_info: Cardinfo;
  }
// ...
  interface Votelist {
    id: string;
    content: string;
    choice_icon: string;
    pic: string;
    pid: string;
    selected: number;
    part_num: string;
    part_ratio: string;
    is_answer: number;
    tick_icon: string;
    content_style: number;
    color?: Color;
  }

  interface Color {
    text_color: string;
    text_color_dark: string;
    border_color: string;
    border_color_dark: string;
    fill_color: string;
    fill_color_dark: string;
    bg_color: string;
    bg_color_dark: string;
    tick_icon: string;
  }
// ...
  interface Continuetag {
    title: string;
    pic: string;
    scheme: string;
  }

  interface Titlesource {
    name: string;
    url: string;
    image: string;
  }

  interface Numberdisplaystrategy {
    apply_scenario_flag: number;
    display_text_min_number: number;
    display_text: string;
  }

  interface Annotation {
    super_tag_id?: string;
    super_exparams?: string;
    shooting?: number;
    client_mblogid?: string;
    source_text?: string;
    phone_id?: string;
    mapi_request?: boolean;
  }

  interface User {
    id: number;
    idstr: string;
    pc_new: number;
    screen_name: string;
    profile_image_url: string;
    profile_url: string;
    verified: boolean;
    verified_type: number;
    domain: string;
    weihao: string;
    verified_type_ext: number;
    status_total_counter: Statustotalcounter;
    avatar_large: string;
    avatar_hd: string;
    follow_me: boolean;
    following: boolean;
    mbrank: number;
    mbtype: number;
    v_plus: number;
    user_ability: number;
    planet_video: boolean;
    icon_list: Iconlist[];
  }

  interface Iconlist {
    type: string;
    data: Data;
  }

  interface Data {
    mbrank: number;
    mbtype: number;
    svip: number;
    vvip: number;
  }

  interface Statustotalcounter {
    total_cnt_format: string;
    comment_cnt: string;
    repost_cnt: string;
    like_cnt: string;
    total_cnt: string;
  }

  interface Visible {
    type: number;
    list_id: number;
  }

  interface BasicItem {
    Link: string;
    Author: string;
    Chapter: string;
  }

  interface WithVoteData extends BasicItem {
    VoteList: Votelist[];
  }
}